requirements.txt
djangorestframework==3.13.1
django-filter==22.1
django_extensions==3.2.0
Pillow==7.0.0
dj-database-url==0.5.0
Django==4.0.6
gunicorn==20.1.0
Pygments==2.0.2
Markdown==3.1.1
whitenoise==6.2.0
django-storages==1.6.5
mysqlclient==2.1.1
matplotlib==3.5.1
